Move DictChunkStore to its own file
The base ChunkStore class got too big for chunkstore.py...
