+++ katdal/test/test_chunkstore.py
@@ -17,17 +17,12 @@
 """Tests for :py:mod:`katdal.chunkstore`."""
 
 import numpy as np
-from numpy.testing import assert_array_equal
 from nose.tools import assert_raises
 
-from katdal.chunkstore import ChunkStore, DictChunkStore
+from katdal.chunkstore import ChunkStore
 
 
-class TestDictChunkStore(object):
-    def setup(self):
-        self.x = np.arange(10)
-        self.y = np.arange(24.).reshape(4, 3, 2)
-        self.store = DictChunkStore(x=self.x, y=self.y)
+class TestChunkStore(object):
 
     def test_store(self):
         store = ChunkStore()
@@ -47,25 +42,3 @@ class TestDictChunkStore(object):
                       chunk=np.array(10 * [{}]))
         assert_raises(ValueError, store.chunk_metadata, "x", [slice(0, 2)],
                       dtype=np.dtype(np.object))
-
-    def test_get(self):
-        s = (slice(3, 5),)
-        actual = self.store.get('x', s, np.dtype(np.int))
-        desired = self.x[s]
-        assert_array_equal(actual, desired, "Error getting x[%s]" % (s,))
-        s = (slice(1, 4), slice(1, 3), slice(1, 2))
-        actual = self.store.get('y', s, np.dtype(np.float))
-        desired = self.y[s]
-        assert_array_equal(actual, desired, "Error getting y[%s]" % (s,))
-
-    def test_put(self):
-        s = (slice(3, 5),)
-        self.store.put('x', s, np.arange(2))
-        actual = self.x[:5]
-        desired = np.array([0, 1, 2, 0, 1])
-        assert_array_equal(actual, desired, "Error putting x[%s]" % (s,))
-        s = (slice(0, 2), slice(0, 3), slice(0, 2))
-        self.store.put('y', s, np.zeros((2, 3, 2), dtype=np.dtype(np.float)))
-        actual = self.y[:2, :3, :]
-        desired = np.zeros((2, 3, 2))
-        assert_array_equal(actual, desired, "Error putting y[%s]" % (s,))
